pyproject.toml
[tool.poetry]
name = "cloudrun-api"
version = "0.1.0"
description = ""
authors = ["mloh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
fastapi = "^0.101.1"
python-json-logger = "^2.0.7"
pydantic = "^2.1.1"
google-cloud-datastore = "^2.17.0"
google-cloud-pubsub = "^2.18.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
requests = "^2.31.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
